+++ classes/activity_meta.php
@@ -59,112 +59,26 @@ class activity_meta {
 
     // Strings.
     /**
-     * @var string $submittedstr - string to use when submitted.
-     */
-    public $submittedstr;
-
-    /**
-     * @var string $notsubmittedstr - string to use when not submitted
-     */
-    public $notsubmittedstr;
-
-    /**
      * @var string $submitstrkey - language string key.
      */
     public $submitstrkey;
 
-    /**
-     * @var string $draftstr - string for draft status.
-     */
-    public $draftstr;
-
-    /**
-     * @var string $reopenedstr - string for reopened status.
-     */
-    public $reopenedstr;
-
-    /**
-     * @var string $expiredstr - string for expired status.
-     */
-    public $expiredstr;
-
-    /**
-     * @var string $notopenstr - string for not open status.
-     */
-    public $notopenstr;
-
-    // General meta data.
-    /**
-     * @var int $timeopen - unix time stamp for time open.
-     */
-    public $timeopen;
-
-    /**
-     * @var int $timeclose - unix time stamp for time closed.
-     */
-    public $timeclose;
-
-    /**
-     * @var int $extension - unix time stamp for extended due dates.
-     */
-    public $extension;
-
+    // Teacher meta data.
     /**
      * @var bool $isteacher - true if meta data is intended for teacher.
      */
     public $isteacher = false;
+
     /**
      * @var bool $submissionnotrequired - true if a submission is not required.
      */
     public $submissionnotrequired = false;
-
-    // Student meta data.
-    /**
-     * @var bool $submitted - true if submission has been made.
-     */
-    public $submitted = false; // Consider collapsing this variable + draft variable into one 'status' variable?
-
-    /**
-     * @var bool $draft - true if activity submission is in draft status.
-     */
-    public $draft = false;
-
-    /**
-     * @var bool $reopened - true if reopened.
-     */
-    public $reopened = false;
-
-    /**
-     * @var bool $expired - true if expired.
-     */
-    public $expired = false;
-
-    /**
-     * @var bool $notopen - true if not open.
-     */
-    public $notopen = false;
-
-    /**
-     * @var bool $notattempted - true if not attempted.
-     */
-    public $notattempted = false;
-
-    /**
-     * @var int $timesubmitted - unix time stamp for time submitted.
-     */
-    public $timesubmitted;
 
     /**
      * @var bool $grade - has the submission been graded.
      */
     public $grade = false;
 
-    /**
-     * @var bool $overdue - is the submission overdue.
-     */
-    public $overdue = false;
-
-    // Teacher meta data.
     /**
      * @var int $numsubmissions - number of submissions.
      */