+++ Changes.md
@@ -13,6 +13,7 @@ Version 3.11.0.1 - TBD
   9. No longer need 'config.php' because of #100.
  10. New 'Accessible back to main course link' - #66.
  11. Apply MDL-71634.
+ 12. Removed conflicting activity meta information that is now in core.
 
 Version 3.10.1.1 - 07/03/2021
 -----------------------------